+++ Kernel/Arch/x86/Hypervisor/VMWareBackdoor.cpp
@@ -7,10 +7,10 @@
 #include <AK/OwnPtr.h>
 #include <AK/Singleton.h>
 #include <Kernel/API/MousePacket.h>
-#include <Kernel/Arch/InterruptDisabler.h>
 #include <Kernel/Arch/x86/Hypervisor/VMWareBackdoor.h>
 #include <Kernel/CommandLine.h>
 #include <Kernel/Debug.h>
+#include <Kernel/InterruptDisabler.h>
 #include <Kernel/Sections.h>
 
 namespace Kernel {
