+++ Kernel/Arch/x86/Time/HPETComparator.cpp
@@ -4,10 +4,10 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-#include <Kernel/Arch/InterruptDisabler.h>
 #include <Kernel/Arch/x86/Time/HPETComparator.h>
 #include <Kernel/Assertions.h>
 #include <Kernel/Debug.h>
+#include <Kernel/InterruptDisabler.h>
 #include <Kernel/Sections.h>
 #include <Kernel/Time/TimeManagement.h>
