+++ Kernel/Arch/x86/Time/RTC.cpp
@@ -4,11 +4,11 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-#include <Kernel/Arch/InterruptDisabler.h>
 #include <Kernel/Arch/x86/IO.h>
 #include <Kernel/Arch/x86/NonMaskableInterruptDisabler.h>
 #include <Kernel/Arch/x86/Time/RTC.h>
 #include <Kernel/Arch/x86/common/CMOS.h>
+#include <Kernel/InterruptDisabler.h>
 #include <Kernel/Time/TimeManagement.h>
 
 namespace Kernel {
