+++ Kernel/Arch/x86/common/InterruptManagement.cpp
@@ -6,7 +6,6 @@
 
 #include <AK/ByteReader.h>
 #include <Kernel/API/Syscall.h>
-#include <Kernel/Arch/InterruptDisabler.h>
 #include <Kernel/Arch/Interrupts.h>
 #include <Kernel/Arch/x86/InterruptManagement.h>
 #include <Kernel/Arch/x86/common/Interrupts/APIC.h>
@@ -14,6 +13,7 @@
 #include <Kernel/Arch/x86/common/Interrupts/PIC.h>
 #include <Kernel/CommandLine.h>
 #include <Kernel/Firmware/MultiProcessor/Parser.h>
+#include <Kernel/InterruptDisabler.h>
 #include <Kernel/Interrupts/SharedIRQHandler.h>
 #include <Kernel/Interrupts/SpuriousInterruptHandler.h>
 #include <Kernel/Memory/TypedMapping.h>
