+++ Kernel/FileSystem/FileSystem.cpp
@@ -7,9 +7,9 @@
 #include <AK/HashMap.h>
 #include <AK/Singleton.h>
 #include <AK/StringView.h>
-#include <Kernel/Arch/InterruptDisabler.h>
 #include <Kernel/FileSystem/FileSystem.h>
 #include <Kernel/FileSystem/Inode.h>
+#include <Kernel/InterruptDisabler.h>
 #include <Kernel/Memory/MemoryManager.h>
 #include <Kernel/Net/LocalSocket.h>
