+++ Kernel/GlobalProcessExposed.cpp
@@ -8,7 +8,6 @@
 #include <AK/JsonObjectSerializer.h>
 #include <AK/Try.h>
 #include <AK/UBSanitizer.h>
-#include <Kernel/Arch/InterruptDisabler.h>
 #include <Kernel/Arch/InterruptManagement.h>
 #include <Kernel/Arch/x86/ProcessorInfo.h>
 #include <Kernel/Bus/PCI/API.h>
@@ -20,6 +19,7 @@
 #include <Kernel/FileSystem/FileBackedFileSystem.h>
 #include <Kernel/FileSystem/OpenFileDescription.h>
 #include <Kernel/Heap/kmalloc.h>
+#include <Kernel/InterruptDisabler.h>
 #include <Kernel/Interrupts/GenericInterruptHandler.h>
 #include <Kernel/KBufferBuilder.h>
 #include <Kernel/Net/LocalSocket.h>
