+++ Kernel/Net/NetworkAdapter.cpp
@@ -4,8 +4,8 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-#include <Kernel/Arch/InterruptDisabler.h>
 #include <Kernel/Heap/kmalloc.h>
+#include <Kernel/InterruptDisabler.h>
 #include <Kernel/Net/EtherType.h>
 #include <Kernel/Net/NetworkAdapter.h>
 #include <Kernel/Net/NetworkingManagement.h>
